requirements.txt
# --- Core LangChain Framework ---
langchain==0.3.7
langchain-core>=0.3.35,<0.4.0
langchain-community==0.3.5
langchain-pinecone==0.2.8

# --- HuggingFace Integration ---
langchain-huggingface
transformers
huggingface-hub

# --- Embeddings ---
sentence-transformers==4.1.0

# --- Data Handling ---
pypdf==5.6.1

# --- Environment ---
python-dotenv==1.1.0

# --- API Framework ---
fastapi==0.115.0
uvicorn==0.30.1


